Hiệu chỉnh lại AI OCR và chỉnh lại Topup có thể dùng hook và có Fallback cho FE

+++ Application/Models/OCR/OCRRegion.cs
@@ -6,10 +6,9 @@ namespace Application.Models.OCR
     {
         public string Text { get; set; } = string.Empty;
         
-        // Coordinates (pixels initially, frontend/API might convert to percentages)
-        public int X { get; set; }
-        public int Y { get; set; }
-        public int Width { get; set; }
-        public int Height { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
+        public double Width { get; set; }
+        public double Height { get; set; }
     }
 }
